backend/requirements.txt
fastapi
uvicorn
python-multipart
numpy-stl
paho-mqtt
requests
sqlalchemy
pydantic
pydantic-settings
python-dotenv
aiofiles
